refactor: db interaction separated from sortinghat
Database methods are passed as arguments to SortingHat class for easier
testing and cleaner architecture.

Moved to_name method of SortingHat to the SortingHatStep class, so that
the SortingHat model only has methods to perform each operation
separately, but the entity that ties them all together is the Step. This
makes it easier to understand what the step does without deep diving
into the SortingHat logic, which now only handles the assignment of the
alerce ID.